+++ CHANGELOG.md
@@ -7,6 +7,12 @@ All notable changes to this project will be documented in this file.
 ### Enhancements
 
 - ASE interface for DFTB+ calculations added
+- init_velocities keyword is ignored if non-zero velocities are present
+- init_velocities can now be forced via the 'force' option
+
+### Bug Fixes
+
+- Volume now gets correctly printed to the .log output file
 
 ### CI